Route native selection through checked instruction and stack effects

+++ src/mir65816/emit/proof.rs
@@ -14,7 +14,7 @@ pub struct Snapshot {
     pub overflow: Option<bool>,
     pub m: Width,
     pub index: Width,
-    pub depth: i32,
+    pub depth: i64,
     pub homes: Vec<(u16, u8, Value)>,
 }
 impl Snapshot {
